Use 32-bit writes exclusively to access clicintx regs
This fixes an issue with unaligned writes.

+++ riscv-peripheral/src/clic/intctl.rs
@@ -4,7 +4,7 @@
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 #[repr(transparent)]
 pub struct INTCTL {
-    ptr: *mut u8,
+    ptr: *mut u32,
 }
 
 impl INTCTL {
